@@ -1,12 +1,14 @@
 module tophatdemon.com/total-invasion-ii
 
-go 1.21
+go 1.22
 
 require (
 	github.com/go-gl/gl v0.0.0-20211210172815-726fda9656d6
 	github.com/go-gl/glfw/v3.3/glfw v0.0.0-20221017161538-93cebf72946b
 	github.com/go-gl/mathgl v1.0.0
 )
+
+require github.com/Southclaws/opt v0.6.1 // indirect
 
 require (
 	github.com/go-audio/audio v1.0.0 // indirect
